Balance calculation now based on double-entry accounting method
The commit changes the model of keeping balances from a single list of
UserBalanceEntry records to a model based on double-entry accounting
principles. It is still to be properly implemented (does not do
verification of all entries and is only keeping track of internal user
accounting books) but we can determine a balance now by the checking
all credits and debits.

This model also tries to ensure that all credits and debits can be
traced back to one recorded model on the database. This can make it
easier for us to calculate balances without relying on state and make
it purely functional. For example, transfers generate a
TransferExecution and TransferConfirmation records. The sender of the
transfer will have the debit associated with the execution record. In
case of a blockchain transfer that fails (or a chain re-org), the
associated transaction will be deleted and cascade to the execution
record and the debit as well.

This approach also makes tests more functional and easier to
reason. To get a desired test condition, all it takes is to call a
factory that generates the data records.