(feat) Deposit model
When developing the UI for the funding section from the web
application, it was possible to realize that the deposit process
should not require the user to specify an amount beforehand. The user
should need to only specify what token it wants to make a deposit so
that hub20 can generate the proper routes, and then any confirmed payment made
through any of the routes should be enough to consider the deposit
amount to be added to the user's balance.