(feat) add listeners for (raiden) channel deposits and token mint
Not all value transfers are based solely on ERC20 `Transfer` events,
so we need to implement listeners for those. This commit adds the
ERC223 ABI, which extends the ERC20 interface to add methods for token
burn and minting and we can then process the `Minted` events.

Related to Raiden, we also now process the `ChannelNewDeposit` event
so that we can register in the accounting books funds transfering from
Layer-1 to Layer-2