(feat) Dockerfile and script to run raiden intergated with Hub20
"Standard" Raiden assumes that at least the keystore file is present and
that is password protected. We want to make it possible to start raiden
based on secrets and environments variables provided by the
docker-compose configuration to run Hub20 instance.

To solve this, we create a custom Dockerfile that takes the private key
as a secret from the docker setup, generates the keystore file (with a
different random, strong password on each start) and then starts raiden.