(chore) Simplify configuration for serving swagger UI
Instead of having near-identical url modules and being forced to run
two separate applications when the user wants to serve the SwaggerUI
page, it is better to make the base url module configurable.

This commit includes a setting that is controlled by *either* the
`HUB20_SERVE_OPENAPI_URLS` or the `HUB20_DEBUG` environment variables. The
endpoints "/docs" and "/openapi" will be included in the list of
provided urlpatterns if any of envvars variables is set.