(fix) Refactor method to listen to incoming transfers
The existing method to listen to incoming transfers grows
geometrically with the number of tokens and accounts being tracked and
is horribly inefficient.

These have been removed in favor an event listener that checks *only*
for deposits with open blockchain routes.

Still need to implement something similar for outgoing
transfers (payouts)