(feat) Introduce celery_pubsub, prepare work for multi-chain
To allow for parallel processing of the many event handlers that may
be triggered by any new event, we use a publish-subscribe pattern that
offloads all the processing from the main event stream thread into
celery tasks

Because now the event stream process is simplified, this opens the way
to allow multi-chain system. We add the `enabled` attribute to the
Chain model in order to make it easy for hub operators to switch on
and off the chains that they want to adopt.