(feat) improved management of blockchain routes.
If a blockchain payment route has already confirmed payments done, it
can be freed for the next deposit. We can also re-use the destination
account for two different deposits if each deposit will be tracking a
different chain and/or token. This reduces the amount of account
generation massively and prevents unnecessary dispersion of funds
around multiple accounts.