(feat) Gas Price Estimate as part of network status. Updated API.
In order for the clients to be able to estimate the transaction costs,
we are adding the estimated gas price as calculated by web3's default
'fast' strategy. We also need to provide an estimate for the cost of a
Raiden deposit, so the API endpoint now also includes information
about raiden.

As part of the process to provide support for multiple chain
simultaneously, we also updated the "/network/status" endpoint. Now it
provides a list of status data, each entry for one specific
blockchain, which can be determined by the "network_id" field. Another
endpoint ("/status/networks/<chain_id>") was also added to provide
information for a particular blockchain.

Also as part of the effort to support multi-chain, the "Chain.make"
method was removed, as it was assuming by default that the
provider_url attribute would come from the settings. Now one must
explicitly call "Chain.objects.get" to construct a chain object.