(feat) create Web3Provider and Explorer model
In order to decouple the information about chains from the web3 rpc
nodes that we connect, a new model needs to be created.

We also create a model for storing information about blockchain
explorer, as a lot of these web3 projects provide one already and we
can use that to present the links to the user.