(chore) break wallet models out of ethereum_money app
The different wallet models that we are using are not related to the
`ethereum_money` package. Considering that we will need to make big
changes to this "money" application in the future (as we eventually
will need to keep track of fiat prices and the relationship of tokens
across chains), it will be better to break unrelated functionality
sooner than later.