Make Raiden configuration possible through Admin
Setting up the Raiden integration correctly requires *both* the raiden
node URL and a working web3 provider so that we can know which chain
the raiden node is connected to. This commit makes it possible for hub
operators to configure the raiden URL through the admin interface.

Once a raiden node is setup properly, the Hub instance will be able to
get information from the payments and record events in the token
networks.