add model manager for double entry accounting models
The manager adds some convenient query methods that allows to take a
look at the models with their balances, as well as to find the models
that have the balance for some funds. This can be used, e.g, whenever we
need to ensure that a wallet has funds to process a transfer.