separate transaction data from transaction receipts into two tables
When we store transactions, the thing that is really important to
us is the receipt, because transaction data records by themselves are
not indicative that the transaction has been mined. Still, it might be
interesting for us to store transaction data entries, as we can use
those records to check if there are transactions that we are waiting
to be mined.

So, this commit creates a separate "TransactionDataRecord" table,
which allows us to store those records, and we now ensure that the
"Transaction" model is concerned only with data that is available in a
transaction receipt.