refactor event streams handlers.
Summary of changes:

Introduce `EventIndexer` table, which keeps track of what blocks have
been processed by each individual stream handler

Ensure that no stream handler makes extra queries for transaction data
or receipts. The basic "process mined blocks" stream now just gets the list
of transactions, and it is up to the handlers to determine if the a
transaction needs to be tracked or not.

Besides the "process mined blocks" stream, ensure that any other
handler is just processing contract events. At the moment, we have one
to collect all `Transfer` events (related to any token transfer between two
addresses) and one to collect `ChannelOpened` and `ChannelClosed`
events from raiden token networks.

This approach reduces *significantly* the amount of requests that are
made to the web3 providers. The downside is that it makes it more
difficult to unit test the stream handler functions.