Refactor relationship betweeen raiden <-> wallet (account) model
With the idea that a raiden can be deployed on multiple chains, we
needed to revisit the concept of having Raiden as a type of ethereum
account, given that the same address can have raiden nodes connected
to different chains.

A Raiden node is now represented only by its URL, the underlying
account address and the chain that is connected to. Given that there
is not yet a reliable way to [get chain information from Raiden's
API](https://github.com/raiden-network/raiden/issues/7483), the
operator needs to include this information when creating the
entry.