improved transfer data handling functions
Use the event_streams functions provided by blockchain and token
modules mainly for indexing, while the "client integrations" handlers
on the core app should be for the real-time operations.

To illustrate, the blockchain modules will be indexing all
transactions related to accounts and the ethereum money app will be
indexing all transfer events. The core app would however only focus on
the listening for events related to curenntly open deposits