Split asyncio-based event handler into multiple docker services
There is no real need for the current event handlers to be
asynchronous. This commit turns them into synchronous functions which
can be called by a django management command. The docker compose file
was then changed to make one service for each event processor.

The existing `run_event_streams` command was renamed to
`run_async_stream_processors` and was kept just in case some
external user wants to add another function that requires asyncio, but
it is not really required.

Closes #214 and #215