Turn stream processor functions into celery beat tasks
Due to some weird buffering issue with docker(-compose?), the django
commands that rely on sleep are causing some timeouts when bringing
services up or down.

Considering that these functions are now all synchronous and can be
executed by celery, it would be easier/better to refactor all of it
out of the single management command and let them be managed by
celery.

To avoid the possibility of multiple workers trying to execute the
same task, we are adding a lock mechanism that can be used by these
tasks that deal with chain querying and event logging.