(feat) chain metadata model
We can use some of the information about the different blockchains to
alleviate the issues arising with the differences in the RPC APIs.
This commit adds a ChainMetadata model which provides fields to
indicate that the chain is a testnet/sidechain/rollup.

A json file with some of the metadata is also added. Ideally this
metadata could become some sort of collaborative effort - by extending
or getting incorporated into chainlist.org for example.