improvements to celery task handling
(fix) reduce number of database queries when checking transfer logs

The method that published the "token_transfer.mined" message to the
subscriber used to expect the handlers to find out the relevant
account addresses. This might give some flexibility, but it also
forces all these different functions to make a new database query on
every transfer event.

Given that we are only interested in token transfer events that are
related to any of the accounts managed by the hub, we check if the
event beforehand and only send the message in case it is related to
any account.

(feat) proper implementation of lock for web3 stream processor tasks

The lock can be refreshed by the task, which resets the timeout. This
way, the task can hold the lock for as long as needed and it still
expires in case of errors.