Move accounting handlers to the proper web3/raiden apps
Create PAYMENT_NETWORK_NAME as app parameter that wants to operate as
django app that implements integration with Hub20 core app. This needs
to be used so that one `PaymentNetworkAccount` model can be registered
for each network and allow its one accounting book

TODO: maybe this can be done directly as part of the app
configuration/setup process?