(feat) payment processing based on specific django command
This commit removes the celery-based approach to collect all blocks,
and it breaks down the intended functionality in two separate
processes/commands:

  - `run_payment_providers` is a command where the state of the
  payment network (e.g, blockchain/raiden node) is pulled into the
  internal database. Each active network provider is place on a
  separate thread and is expected to implement an infinite loop.

  - `run_payment_processors` is a command that is meant to check for
  all open payment routes and place it in a separate command, where
  the route can execute its own `process` command. In the case of a
  blockchain payment, this means that the route checks for payments
  made to allocated wallet until a payment is made or the route is
  expired (i.e, the chain block height is beyond the payment window)

The separation of these commands comes from the realization that we do
not really need to bring all the history from the blockchain to our
database in order to be able to accept payments, or even to make
outside transfers. We might eventually be interested in implementing
proper blockchain indexers or (more plausible) simply using the
existing blockchain explorers to collect historical data related to
our wallet history.